@@ -2,7 +2,7 @@ package ru.fmproject.android.calculator;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.v4.widget.DrawerLayout;
+import androidx.drawerlayout.widget.DrawerLayout;
 import android.util.AttributeSet;
 import android.view.Gravity;
 import android.view.View;
